backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.12.0